requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
urllib3==2.2.1
